refactor version endpoint into generic static endpoint
add version function moved and renamed into add static data,
which can be used to serve arbitrary data that is static
and can be represented as json
